Add files via upload
In your choice of language(Preferably python/R), write a program that prints the numbers ranging from one to 100. But for multiples of three, print "Fizz" instead of the number, and for the multiples of five, print "Buzz." For numbers which are multiples of both three and five, print "FizzBuzz".